src/requirements.txt
pydantic==2.10.6
python-dotenv==1.0.1
inquirer==3.4.0

